bcplus-master/text_analysis/requirements.txt
asgiref==3.3.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-language==2.0.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
JPype1==1.1.2
konlpy==0.5.2
libcst==0.3.13
lxml==4.6.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pandas==1.1.4
proto-plus==1.11.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.11
webencodings==0.5.1
wincertstore==0.2
django-cors-headers==3.5.0
